requirements.txt
streamlit==1.30.0
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
langchain==0.1.164
langchain-google-genai==0.1.11
faiss-cpu==1.7.4
python-dotenv==1.0.1
